Update après les merges des différentes pull request

+++ formation_metier/views/inscription_formation_pour_participant_view.py
@@ -11,6 +11,7 @@ from django.views.generic.detail import SingleObjectMixin, DetailView
 
 from formation_metier.models.formation import Formation
 from formation_metier.models.inscription import Inscription
+from formation_metier.models.seance import Seance
 
 
 class InscriptionFormationPourParticipantForm(forms.ModelForm):
@@ -20,7 +21,7 @@ class InscriptionFormationPourParticipantForm(forms.ModelForm):
 
 
 class InscriptionFormationPourParticipant(LoginRequiredMixin, DetailView, SingleObjectMixin):
-    model = Inscription
+    model = Formation
     pk_url_kwarg = 'formation_id'
     context_object_name = "formation"
     template_name = "formation_metier/inscription_formation_pour_participant.html"
@@ -92,11 +93,12 @@ class InscriptionFormationPourParticipant(LoginRequiredMixin, DetailView, Single
 
     def delete(self, request, inscription_list):
         for inscription in inscription_list:
-            inscription.delete()
-            messages.success(
-                request,
-                f"Votre inscription pour la seance du {inscription.seance.datetime_format()} a été supprimée"
-            )
+            if inscription.seance.seance_date > datetime.now():
+                inscription.delete()
+                messages.success(
+                    request,
+                    f"Votre inscription pour la seance du {inscription.seance.datetime_format()} a été supprimée"
+                )
 
     def create(self, request, seances_liste):
         for seance_id in seances_liste:
@@ -104,25 +106,7 @@ class InscriptionFormationPourParticipant(LoginRequiredMixin, DetailView, Single
                 participant=request.user.employeuclouvain,
                 seance_id=seance_id
             )
-            seance_list_apres_post = self.request.POST.getlist('seance')
-            for inscription_existante in inscriptions_existantes_avant_post:
-                if str(inscription_existante.seance.id) not in seance_list_apres_post and inscription_existante.seance.seance_date > datetime.datetime.now():
-                    inscription_existante.delete()
-                    messages.success(
-                        request,
-                        f"Votre inscription pour la seance du {inscription_existante.seance.datetime_format()} a été supprimée"
-                    )
-            for seance_id in seance_list_apres_post:
-                if not Inscription.objects.filter(participant__user=request.user, seance_id=seance_id).exists():
-                    seance_object = Seance.objects.get(id=seance_id)
-                    inscription_cree = Inscription.objects.create(
-                        participant=request.user.employeuclouvain,
-                        seance=seance_object
-                    )
-                    messages.success(
-                        request,
-                        f"Votre inscription pour la seance du {inscription_cree.seance.datetime_format()} a été sauvegardée"
-                    )
-            return redirect(
-                self.get_success_url()
+            messages.success(
+                request,
+                f"Votre inscription pour la seance du {inscription_cree.seance.datetime_format()} a été sauvegardée"
             )
